requirements.txt
apache-airflow-providers-google==10.5.0